requirements/dev.txt
-r base.txt

# developer specific dependencies
# used for development, testing, and CI

# debugging
pudb==2014.1

# testing
coverage==3.6
mock==1.0.1
django-autofixture>=0.8.0
python-coveralls
django-discover-runner==1.0
django-nose==1.4

# developer tools
django-debug-toolbar==1.3
django-anonymizer==0.5
django-extensions==1.5.5
